% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_data_fields.R
\name{daily.data.fields}
\alias{daily.data.fields}
\title{Daily data is obtained from hourly data}
\usage{
daily.data.fields(raster.list, aggregate)
}
\arguments{
\item{raster.list}{Spat Raster variable with several times for a unique variable (T2 or HR2 or ...)}

\item{aggregate}{Type of aggregation (sum, mean, min, max)}
}
\value{
Spat Raster with daily information
}
\description{
....
}
