% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily2monthly.R
\name{\%>\%}
\alias{\%>\%}
\alias{daily2monthly}
\title{Daily data to monthly}
\usage{
daily2monthly(data = data)
}
\arguments{
\item{data}{matrix with daily data from mg.evaluation output function}
}
\value{
Data frame with monthly data
}
\description{
Data transformation from daily to monthly scale
}
