% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva-data.R
\docType{data}
\name{eva}
\alias{eva}
\title{Evaporation data (observation and model)}
\format{
An object of class \code{"data.frame"}.
\describe{
\item{Dates}{1st column with dates}
\item{evapo_obs}{2nd column with evaporation observation}
\item{OUT_PREC}{Precipitation}
\item{OUT_EVAP}{Evaporation}
\item{OUT_RUNOFF}{Runoff}
\item{OUT_BASEFLOW}{Baseflow}
\item{OUT_SOIL_MOIST_lyr_1}{Soil moisture from 1st layer}
\item{OUT_EVAP_CANOP}{Evaporation from canopy}
\item{OUT_SURF_TEMP}{Surface temperature}
}
}
\usage{
data(eva)
}
\description{
Data of evaporation from in-situ observation and several soil model outputs
}
\examples{
data(eva)
}
\references{
Diaz et al. (2024) AAGG 2024
Not yet published
}
\keyword{datasets}
