% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nearest_point.R
\name{find.nearest.point}
\alias{find.nearest.point}
\title{Title}
\usage{
find.nearest.point(data.spat.raster = data.spat.raster, lon = lon, lat = lat)
}
\arguments{
\item{data.spat.raster}{Spat Raster of WRF SMN (only one or several)}

\item{lon}{Longitude location of nearest point to find}

\item{lat}{Latitude location of nearest point to find}
}
\value{
a vector with the nearest location (lon/lat) and time serie of that location
}
\description{
Location of nearest point to lon/lat and temporal serie of location
}
