% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wrf_files.R
\name{get.wrf.files}
\alias{get.wrf.files}
\title{Title}
\usage{
get.wrf.files(anual = anual, mes = mes, dia = dia, ciclo = ciclo, time = time)
}
\arguments{
\item{anual}{character or numeric of year}

\item{mes}{character or numeric of month}

\item{dia}{character or numeric of day}

\item{ciclo}{cycle of forecast, "00", "06", "12" or "18"}

\item{time}{selection of datasets, "01H", "24H" or "10M"}
}
\value{
string of the list of elements in the Bucket
}
\description{
Character string with the filenames of WRF SMN located in AWS Bucket
}
