% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_by_variable.R
\name{load.by.variable}
\alias{load.by.variable}
\title{Title}
\usage{
load.by.variable(
  nc.filenames = nc.filenames,
  variable = variable,
  transform = transform,
  method = method
)
}
\arguments{
\item{nc.filenames}{netcdf files}

\item{variable}{name of variable from https://odp-aws-smn.github.io/documentation_wrf_det/Formato_de_datos/ as character}

\item{transform}{TRUE to project data to longlat datum=WGS84}

\item{method}{if transform is set TRUE define projection method taken from project function of terra package}
}
\value{
Spat Raster
}
\description{
Open of netcdf files of WRF SMN drom AWS and optional projection
}
