% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_evaluation.R
\name{mg.evaluation}
\alias{mg.evaluation}
\title{Evaluation of regression}
\usage{
mg.evaluation(
  input.data = input.data,
  predictand = predictand,
  predictors = predictors,
  var.model = var.model,
  lmodel = lmodel
)
}
\arguments{
\item{input.data}{Data frame with first column as a "POSIXct" class and one or
more columns with data values. The predictand and predictors variables should
be located in these columns}

\item{predictand}{Character with column name of the predictand variable}

\item{predictors}{Character array with one or more elements of the predictors
chosen by the user}

\item{var.model}{Character with column name of the modeled predicting variable}

\item{lmodel}{Element of class lm obtained from multiple.guidance output
function}
}
\value{
List of two elements. First element is a matrix with the columns of
observed data, modeled data and corrected data. Second element is a data
frame of the statistical results of the modeled and corrected data versus
observed data
}
\description{
Evaluation of the linear multiple regression obtained from the
multiple.guidance function
}
