% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_guidance.R
\name{multiple.guidance}
\alias{multiple.guidance}
\title{Multiple lineal regression of data}
\usage{
multiple.guidance(
  input.data = input.data,
  predictand = predictand,
  predictors = predictors
)
}
\arguments{
\item{input.data}{Data frame with first column as a "POSIXct" class and one or
more columns with data values. The predictand and predictors variables should
be located in these columns}

\item{predictand}{Character with column name of the predictand variable}

\item{predictors}{Character array with one or more elements of the predictors
chosen by the user}
}
\value{
an element of class lm
}
\description{
Definition of linear multiple regression adjustment based on predictor
variables that fit a predicting variable
}
