% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploting.R
\name{ploting}
\alias{ploting}
\title{Plot of data}
\usage{
ploting(data = data)
}
\arguments{
\item{data}{Data frame from daily2monthly output function or any other
temporal series}
}
\value{
ggplot element
}
\description{
Plot of observed, modeled and corrected guidance series
}
