% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_download.R
\name{wrf.download}
\alias{wrf.download}
\title{Title}
\usage{
wrf.download(wrf.name = wrf.name)
}
\arguments{
\item{wrf.name}{list of names to download from get.wrf.files}
}
\value{
downloaded netcdf files
}
\description{
Download of WRF SMN data from AWS
}
