% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wrf_files.R
\name{get.wrf.files}
\alias{get.wrf.files}
\title{List of available files for downloading}
\usage{
get.wrf.files(year = year, month = month, day = day, cycle = cycle, time = time)
}
\arguments{
\item{year}{character or numeric of year}

\item{month}{character or numeric of month}

\item{day}{character or numeric of day}

\item{cycle}{cycle of forecast, "00", "06", "12" or "18"}

\item{time}{selection of datasets, "01H", "24H" or "10M"}
}
\value{
string of the list of elements in the Bucket
}
\description{
Character string with the filenames of WRF SMN located in AWS Bucket
}
