% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_by_variable.R
\name{load.by.variable}
\alias{load.by.variable}
\title{Load and projection of data}
\usage{
load.by.variable(nc.filenames, variable, transform, method)
}
\arguments{
\item{nc.filenames}{netcdf files}

\item{variable}{name of variable from https://odp-aws-smn.github.io/documentation_wrf_det/Formato_de_datos/ as character}

\item{transform}{TRUE to project data to longlat datum=WGS84}

\item{method}{if transform is set TRUE define projection method taken from project function of terra package}
}
\value{
Spat Raster with the chosen variable (optional: with the projection changed)
}
\description{
Open of netcdf files of WRF SMN from AWS and optional projection
}
